refactor epic tracking
