+++ src/components/modules/banners/ausMomentContributionsBanner/AusMomentContributionsBanner.tsx
@@ -4,7 +4,7 @@ import { body, headline, textSans } from '@guardian/src-foundations/typography';
 import { neutral, opinion } from '@guardian/src-foundations/palette';
 import { from, until } from '@guardian/src-foundations/mq';
 import { space } from '@guardian/src-foundations';
-import { setContributionsBannerClosedTimestamp } from './localStorage';
+import { setContributionsBannerClosedTimestamp } from '../localStorage';
 import { BannerTracking, BannerProps } from '../../../../types/BannerTypes';
 import SocialLinks from './SocialLinks';
 import { SvgClose } from '@guardian/src-icons';
