+++ src/components/modules/banners/digitalSubscriptions/DigitalSubscriptionsBanner.stories.tsx
@@ -2,6 +2,7 @@ import React, { ReactElement } from 'react';
 import { DigitalSubscriptionsBanner } from './DigitalSubscriptionsBanner';
 import { withKnobs } from '@storybook/addon-knobs';
 import { StorybookWrapper } from '../../../../utils/StorybookWrapper';
+import { BannerTracking } from '../../../../types/BannerTypes';
 
 export default {
     component: DigitalSubscriptionsBanner,
@@ -9,15 +10,16 @@ export default {
     decorators: [withKnobs],
 };
 
-const tracking = {
+const tracking: BannerTracking = {
     ophanPageId: 'kbluzw2csbf83eaberrr',
-    ophanComponentId: 'SUBSCRIPTIONS_BANNER',
     platformId: 'GUARDIAN_WEB',
     clientName: 'dcr',
     referrerUrl: 'http://localhost:3030/Article',
     abTestName: 'DigitalSubscriptionsBanner',
     abTestVariant: 'control',
     campaignCode: '',
+    componentType: 'ACQUISITIONS_SUBSCRIPTIONS_BANNER',
+    products: [],
 };
 
 export const defaultStory = (): ReactElement => {
