+++ src/components/modules/banners/guardianWeekly/GuardianWeeklyBanner.tsx
@@ -19,7 +19,7 @@ import {
     becomeASubscriberButton,
     linkStyle,
 } from './guardianWeeklyBannerStyles';
-import { BannerProps } from '../BannerTypes';
+import { BannerProps } from '../../../../types/BannerTypes';
 import { setSubscriptionsBannerClosedTimestamp } from '../localStorage';
 
 export const GuardianWeeklyBanner: React.FC<BannerProps> = ({
