+++ src/factories/tracking.ts
@@ -7,7 +7,6 @@ import {
 
 export const pageTracking = Factory.define<EpicPageTracking>(() => ({
     ophanPageId: 'k5nxn0mxg7ytwpkxuwms',
-    ophanComponentId: 'ACQUISITIONS_EPIC',
     platformId: 'GUARDIAN_WEB',
     clientName: 'dcr',
     referrerUrl:
@@ -19,6 +18,7 @@ export const testTracking = Factory.define<EpicTestTracking>(() => ({
     campaignId: 'remote_epic_test',
     abTestName: 'remote_epic_test',
     abTestVariant: 'api',
+    componentType: 'ACQUISITIONS_EPIC',
 }));
 
 export const tracking = Factory.define<EpicTracking>(({ factories }) => ({
