+++ src/server.test.ts
@@ -56,6 +56,8 @@ describe('POST /epic', () => {
             campaignCode:
                 'gdnwb_copts_memco_2020-02-11_enviro_fossil_fuel_r2_Epic__no_article_count_Control',
             campaignId: 'epic_2020-02-11_enviro_fossil_fuel_r2_Epic__no_article_count',
+            componentType: 'ACQUISITIONS_EPIC',
+            products: ['CONTRIBUTION', 'MEMBERSHIP_SUPPORTER'],
         });
     });
